MB-45765. If commit submission failed issue rollback
Reviewed-on: http://review.couchbase.org/c/query/+/151541
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>

+++ datastore/couchbase/txcouchbase.go
@@ -197,6 +197,7 @@ func (s *store) CommitTransaction(stmtAtomicity bool, context datastore.QueryCon
 	}
 	logging.Tracea(func() string { return fmt.Sprintf("=====%v=====Commit end write========", txId) })
 
+	cleanUp := true
 	if transaction != nil {
 		var wg sync.WaitGroup
 
@@ -219,20 +220,25 @@ func (s *store) CommitTransaction(stmtAtomicity bool, context datastore.QueryCon
 			if cerr != nil {
 				err = cerr
 			}
+			txMutations.SetTransaction(nil, nil)
+		} else {
+			// if commit request submission failed, let issue rollback
+			cleanUp = false
 		}
 
 		logging.Tracea(func() string { return fmt.Sprintf("=====%v=====Actual Commit end========", txId) })
 
-		txMutations.SetTransaction(nil, nil)
 	} else {
 		err = gcagent.ErrNoTransaction
 	}
 
-	// Release transaction mutations
-	var memSize int64
-	txMutations.DeleteAll(true, &memSize)
-	txMutations.Recycle()
-	txContext.SetTxMutations(nil)
+	if cleanUp {
+		// Release transaction mutations
+		var memSize int64
+		txMutations.DeleteAll(true, &memSize)
+		txMutations.Recycle()
+		txContext.SetTxMutations(nil)
+	}
 
 	if err != nil {
 		e, c := errorType(err, false)
